feat: created predict endpoint
