+++ Dockerfile
@@ -16,7 +16,9 @@ COPY site/css ./site/css
 COPY site/html ./site/html
 COPY prompts ./server/templates
 COPY server/favicon.ico ./server/favicon.ico
-COPY server/* ./server
+COPY server/generated ./server/generated
+COPY server/ai ./server/ai
+COPY server/*.go ./server
 
 COPY server/go.mod server/go.sum ./server/
 WORKDIR /server
